plugins.handlers: type annotations & subfolder bugfix
In `PyFilePlugin.__init__()`, the `filename` must be wrapped in a list
before passing it to `importlib.util.spec_from_file_location()`, whose
`submodule_search_locations` argument is expected to be a list of `str`.

That +1/-1 patch should be released as a hotfix for 8.0.x, but the rest
of the type changes and doc tweaks can wait for 8.1.