coretasks, url: use bot._url_callbacks
In order to emit warnings, the method Sopel.search_url_callbacks must
not be used internally or through a built-in plugin. As a result, and
until removal in Sopel 9, both coretasks and url are going to access
bot._url_callbacks directly.

Also, the method `bot.call` is now deprecated and will be removed in
Sopel 9.0, following a short deprecation cycle. This method should have
been deprecated since Sopel 8 but was kept in order to keep the old URL
callbacks system working. When it is gone, this method won't be needed
anymore.