bot: round rate limit `time_left` value up, eliminate microseconds
This makes rate limit feedback from the bot more readable.

We could go full-on custom format, but honestly the built-in string
representation of a timedelta is perfectly serviceable if you get rid of
the microseconds portion. (How you get there is pretty ugly, but it's
not AS bad as a whole new helper function in `tools.time`.)