ci: make `v` prefix optional in maintenance branch names
We aren't using the `v` prefix for maintenance branches. What we have
are names like `7.1.x`, `8.0.x`, etc. and CI hasn't been running for 8.0
series maintenance patches like it should.

I already proposed this in the 8.0.2 release-prep PR, but it needs to be
in the master branch too so future maintenance branches will get the new
patterns automatically.