cli.plugins: add `sopel-plugins new` subcommand, a cookiecutter wrapper
Running `sopel-plugins new` bootstraps a new plugin in a subfolder of
`.` or the specified `--directory`, using `gh:sopel-irc/plugin` or the
specified `--template` (URL or filesystem path).

The `cookiecutter` package itself must be installed manually; I feel it
brings along too much dependency baggage to make it a standard part of
every Sopel installation.

Maybe in a future world where we start using libraries like `click`
and/or `rich` for Sopel's own CLI, it would be less of an issue. Along
these lines I'm also mulling the idea of calling the `cookiecutter` CLI
entrypoint as a subprocess, instead, so users can use better approaches
such as managing their CLI helpers with `pipx` etc.