plugins: add plugin callable and job classes
Plugin callables and plugin jobs are functions that go through plugin
decorators then the loader's cleaning process, and that relies on adding
attributes to function. This is not really ideal for type checking
purpose, and it can create all kind of weird behavior.

To prepare its replacement, the module `sopel.plugins.callables`
contains all the necessary tools and classes to switch from a flimsy
hacked functions based system to a more sturdy and type safe class based
system.

Nothing actually changes for the bot so far, as things will change in a
future commit.

Note: the various plugin decorators `require_*` are not handled yet by
this new system.