sopel: refactoring of the plugin decorators
Behold! Plugin's decorators are now creating plugin objects, instead of
adding type unsafe attributes to functions.

This is theorically backward compatible with Sopel 8.0 and below's
system of plugin callable (i.e. functions only).

Various changes:

* decorators from `sopel.plugin` now return plugin object
* `require_*` decorators now use a standard way to declare predicates
  (all predicates must return True for the callable to execute)
* `sopel.plugin.capability` is now a function that returns an instance
  of `sopel.plugins.callables.Capability`, with the same behavior
* module `sopel.loader` is now deprecated and will be removed in
  Sopel 9.0
* rule handlers are type annotated with the new plugin objects

This is minor effects on the bot itself, this is more an enabler for
what's to come, opening a wide arrayd of possibilities for plugin author
in term of customization of plugin callables.