plugins: refactor "is triggerable"
Until now, "is triggerable" was a loosely defined concept, because on on
one hand every job callable is technically triggerable, but on the other
hand the `sopel.loader.is_triggerable` function (now deprecated) treats
URL callback as "non triggerable".

On the new plugin callable class, this is not the case:

* is_triggerable will return True for any callable with at least one
  pattern defined
* is_generic_rule and is_named_rule properties are here for parity with
  the is_url_callack property

This should make things more clear for everyone about what is and what
isn't triggerable.