lint: correct TYPE_CHECKING style
New checks in flake8-type-checking 0.19 started enforcing style rules
about import groups within `if TYPE_CHECKING:` and whitespace before the
block itself.

We are not using 0.19 yet because it has a bug with multi-line imports:

    from foo import (
        bar,
        baz,
        eggs,
    )

    if TYPE_CHECKING:
        ...

This is currently reported as TYPE_CHECKING not having a blank line
before it, but will be fixed soon; upstream has a patch already.

One more-substantial change in this commit moves the TYPE_CHECKING block
of `sopel/plugin.py` to directly follow other imports; this is an edge
case not covered by flake8-import-order's patch and would not be trivial
to handle. It's arguably *wrong* to start assigning constants before ALL
imports (including type-check-only) are done, so I'm happy rearranging.