Makefile: swap `cleancache` / `cleancov`
`make clean` is meant to tidy up after creating a test build, which
includes `make qa` and a call to `python -m build`. It doesn't need to
wipe __pycache__ files.

`allclean` is the target for getting rid of (nearly?) everything that
doesn't come with a fresh clone, so that's the one that should wipe
cached bytecode.