isupport: parse CLIENTTAGDENY to a custom type that understands `*`
This replaces directly passing through the CLIENTTAGDENY value as a
plain string.

Consumers of the CLIENTTAGDENY token can either check whether a tag name
is directly contained in the set, or go through the `is_denied()` method
to apply wildcard logic if appropriate.

The `ClientTagDeny` object's membership tests are case insensitive.