tests: mock IRC server's invite method
When Sopel handles a trigger, it may execute commands in their own
threads. This is the expected behavior. However, when testing Sopel's
various commands, this may result in race conditions.

One use case is the INVITE event handler in the `admin` builtin plugin:
because the handler is executed in a thread, the tests would sometimes
result in a failure that was quite hard to understand at first.

To solve that, a new method on the mock IRC server class has been added:
`invite`. This method will send the appropriate message to the bot, then
wait to join all the running thread (unless told not to).

I also added a new test helper function: `sopel.tests.on_message`.
This function will call bot's `on_message` method then join all running
trigger threads.

Co-authored-by: dgw <dgw@technobabbl.es>