tests: use pathlib instead of legacy py.path
The `ConfigFactory` used the legacy py.path object tmpdir from the
pytest fixture: as the documentation of pytest states that this is not
the preferred method, I switched to the preferred method.

Therefore, `ConfigFactory` now uses a builtin pathlib.Path object as
input, and expect a directory. It will create a subdirectory named
`sopel`  that will be used to store the test configuration file.

This has the double effect of removing an unwated dependency to pytest
in that section of the code: everything pytest related is kept in
`sopel.tests.pytest_plugin` only!

I know this is a breaking change, so I added a not on
`sopel.tests.factories`: plugin authors shouldn't create factory
instances themselves, and rely on pytest's fixtures instead, which hide
this kind of implementation details.